move core functionality to appropriate module and make project multiplatform
