+++ .idea/artifacts/dolphins_core_jvm_0_1_0.xml
@@ -1,0 +1,6 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="dolphins-core-jvm-0.1.0">
+    <output-path>C:\Users\miralace\space\dolphins\core\build\libs</output-path>
+    <root id="archive" name="dolphins-core-jvm-0.1.0.jar" />
+  </artifact>
+</component>
