+++ .idea/artifacts/rxjava_js_0_0_1.xml
@@ -1,0 +1,6 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="rxjava-js-0.0.1">
+    <output-path>C:\Users\miralace\space\dolphins\rxjava\build\libs</output-path>
+    <root id="archive" name="rxjava-js-0.0.1.jar" />
+  </artifact>
+</component>
