+++ .idea/artifacts/rxjava_jvm_0_0_1.xml
@@ -1,0 +1,6 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="rxjava-jvm-0.0.1">
+    <output-path>C:\Users\miralace\space\dolphins\rxjava\build\libs</output-path>
+    <root id="archive" name="rxjava-jvm-0.0.1.jar" />
+  </artifact>
+</component>
