+++ .idea/compiler.xml
@@ -15,5 +15,9 @@
       <entry name="!*.mpsr" />
       <entry name="!*.model" />
     </wildcardResourcePatterns>
+    <bytecodeTargetLevel>
+      <module name="dolphins.main" target="1.8" />
+      <module name="dolphins.test" target="1.8" />
+    </bytecodeTargetLevel>
   </component>
 </project>
