+++ core/src/commonMain/kotlin/dolphins/foundation/Kind.kt
@@ -1,4 +1,4 @@
-package ru.hbracket.dolphins.f
+package dolphins.foundation
 
 interface Kind<out F, out A>
 typealias Kind2<F, A, B> = Kind<F, Kind<A, B>>
