+++ core/src/commonMain/kotlin/dolphins/foundation/ShadyException.kt
@@ -1,4 +1,4 @@
-package ru.hbracket.dolphins.f
+package dolphins.foundation
 
 class ShadyException(className: String) : Exception(
     "The instance of this $className is shady and behaves not according to laws for the sake of performance."
