+++ core/src/commonMain/kotlin/dolphins/foundation/typeclasses/Consume.kt
@@ -1,6 +1,6 @@
-package ru.hbracket.dolphins.f.ts
+package dolphins.foundation.typeclasses
 
-import ru.hbracket.dolphins.f.Kind
+import dolphins.foundation.Kind
 
 interface Consume<F, H : Handle> {
